+++ Sources/API/Model/Endpoints.swift
@@ -38,7 +38,7 @@ enum Endpoints {
     }
     enum Identification {
         enum GetStatus {
-            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<EmptyRequest>, WPNResponse<IdentityStatusResponse>>
+            typealias EndpointType = WPNEndpointAuthenticatedWithToken<WPNRequest<EmptyRequest>, WPNResponse<IdentityStatusResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/status", tokenName: "possession_universal") }
         }
         enum StartReVerification<TRequest: Encodable> {
@@ -46,43 +46,43 @@ enum Endpoints {
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/onboarding/start", uriId: "/api/onboarding/start", e2ee: .activationScope) }
         }
         enum Init {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<IdentityInitRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<IdentityInitRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/init", uriId: "/api/identity/init") }
         }
         enum Cancel {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<ProcessRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<ProcessRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/cleanup", uriId: "/api/identity/cleanup") }
         }
         enum ConsentText {
-            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<ConsentTextRequest>, WPNResponse<ConsentTextResponse>>
+            typealias EndpointType = WPNEndpointAuthenticatedWithToken<WPNRequest<ConsentTextRequest>, WPNResponse<ConsentTextResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/consent/text", tokenName: "possession_universal") }
         }
         enum ConsentApprove {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<ConsentApproveRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<ConsentApproveRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/consent/approve", uriId: "/api/identity/consent/approve") }
         }
         enum DocumentScanSdkInit {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<SDKInitRequest>, WPNResponse<SDKInitResponse>>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<SDKInitRequest>, WPNResponse<SDKInitResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/document/init-sdk", uriId: "/api/identity/document/init-sdk", e2ee: .activationScope) }
         }
         enum SubmitDocumentsV2 {
-            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<DocumentSubmitRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticatedWithToken<WPNRequest<DocumentSubmitRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/v2/identity/document/submit", tokenName: "possession_universal", e2ee: .activationScope) }
         }
         enum DocumentsStatus {
-            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<ProcessRequest>, WPNResponse<DocumentStatusResponse>>
+            typealias EndpointType = WPNEndpointAuthenticatedWithToken<WPNRequest<ProcessRequest>, WPNResponse<DocumentStatusResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/document/status", tokenName: "possession_universal") }
         }
         enum PresenceCheckInit {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<ProcessRequest>, WPNResponse<PresenceCheckInitResponse>>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<ProcessRequest>, WPNResponse<PresenceCheckInitResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/presence-check/init", uriId: "/api/identity/presence-check/init", e2ee: .activationScope) }
         }
         enum PresenceCheckSubmit {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<ProcessRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<ProcessRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/presence-check/submit", uriId: "/api/identity/presence-check/submit") }
         }
         enum ResendOTP {
-            typealias EndpointType = WPNEndpointSigned<WPNRequest<ProcessRequest>, WPNResponseBase>
+            typealias EndpointType = WPNEndpointAuthenticated<WPNRequest<ProcessRequest>, WPNResponseBase>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/otp/resend", uriId: "/api/identity/otp/resend") }
         }
         enum VerifyOTP {
@@ -90,7 +90,7 @@ enum Endpoints {
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/otp/verify", e2ee: .activationScope) }
         }
         enum ActivationFinish {
-            typealias EndpointType = WPNEndpointSignedWithToken<WPNRequest<ActivationFinishRequest>, WPNResponse<ActivationFinishResponse>>
+            typealias EndpointType = WPNEndpointAuthenticatedWithToken<WPNRequest<ActivationFinishRequest>, WPNResponse<ActivationFinishResponse>>
             static var endpoint: EndpointType { .init(endpointURLPath: "/api/identity/activation", tokenName: "possession_universal", e2ee: .activationScope) }
         }
     }