+++ WultraDigitalOnboardingTests/HelperClasses.swift
@@ -34,14 +34,17 @@ class TestHelper {
     private let environment: ServerEnvironment
     
     init(environment: ServerEnvironment, processType: String, customPaInstance: PowerAuthSDK? = nil) throws {
-        guard let pa = customPaInstance ?? PowerAuthSDK(
-            configuration: .init(
-                instanceId: UUID().uuidString,
-                baseEndpointUrl: environment.esUrl,
-                configuration: environment.mobileConfig
+        let pa: PowerAuthSDK
+        if let customPaInstance {
+            pa = customPaInstance
+        } else {
+            pa = try PowerAuthSDK(
+                configuration: .init(
+                    instanceId: UUID().uuidString,
+                    baseEndpointUrl: environment.esUrl,
+                    configuration: environment.mobileConfig
+                )
             )
-        ) else {
-            throw SimpleError("Failed to create PowerAuthSDK")
         }
         
         self.powerAuth = pa
@@ -75,8 +78,8 @@ class TestHelper {
         let result = try await activation.activate(otp: otp, activationName: UIDevice.current.name)
         print("activated: \(result.activationFingerprint)")
         
-        // perist with random password
-        try powerAuth.persist()
+        // persist with random password
+        try await powerAuth.persist()
         
         // verify powerauth status
         let paStatus = try await powerAuth.fetchActivationStatus()
@@ -257,8 +260,16 @@ extension PowerAuthSDK {
         }
     }
     
-    func persist(password: String = UUID().uuidString) throws {
-        return try persistActivation(withPassword: password)
+    func persist(password: String = UUID().uuidString) async throws {
+        try await withCheckedThrowingContinuation { (cont: CheckedContinuation<Void, Error>) in
+            persistActivation(withPassword: password) { error in
+                if let error {
+                    cont.resume(throwing: error)
+                } else {
+                    cont.resume(returning: ())
+                }
+            }
+        }
     }
 }